# app/Http/Controllers/AuthController.php
<?php

namespace App\Http\Controllers;

use App\Core\Auth\Application\Dto\Request\LoginRequest;
use App\Core\Auth\Application\Dto\Request\RegisterRequest;
use App\Core\Auth\Application\Exception\InvalidCredentialsException;
use App\Core\Auth\Application\UseCase\Login;
use App\Core\Auth\Application\UseCase\Logout;
use App\Core\Auth\Application\UseCase\Me;
use App\Core\Auth\Application\UseCase\RefreshToken;
use App\Core\Auth\Application\UseCase\Register;
use App\Core\Home\House\Model\Service\HouseCreator;
use Exception;
use Illuminate\Http\JsonResponse;
use OpenApi\Annotations as OA;

final class AuthController extends Controller
{
    public function __construct(
        private Login        $login,
        private Register     $register,
        private HouseCreator $houseCreator,
        private Logout       $logout,
        private RefreshToken $refreshToken,
        private Me           $me,
    )
    {
    }

    /**
     * @OA\Post(
     *     path="/auth/login",
     *     tags={"Auth"},
     *     summary="Start session",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"email","password"},
     *             @OA\Property(property="email", type="string", format="email", example="user@example.com"),
     *             @OA\Property(property="password", type="string", format="password", minLength=8)
     *         )
     *     ),
     *     @OA\Response(response=200, description="OK", @OA\JsonContent(ref="#/components/schemas/AuthResponse")),
     *     @OA\Response(response=401, description="Unauthorized", @OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function login(LoginRequest $request): JsonResponse
    {
        try {
            $response = $this->login->execute($request);
            return response()->json($response->toArray());
        } catch (InvalidCredentialsException $e) {
            return response()->json(['message' => $e->getMessage()], 401);
        }
    }

    /**
     * @OA\Post(
     *     path="/auth/register",
     *     tags={"Auth"},
     *     summary="Register new user",
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"name","email","password"},
     *             @OA\Property(property="name", type="string", maxLength=255, example="Maria"),
     *             @OA\Property(property="email", type="string", format="email", example="maria@example.com"),
     *             @OA\Property(property="password", type="string", format="password", minLength=5)
     *         )
     *     ),
     *     @OA\Response(response=201, description="Created", @OA\JsonContent(ref="#/components/schemas/AuthResponse")),
     *     @OA\Response(response=401, description="Unauthorized", @OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function register(RegisterRequest $request): JsonResponse
    {
        try {
            $response = $this->register->execute($request);
            $this->houseCreator->create($response->user['id'], "{$response->user['name']}s House");
            return response()->json($response->toArray(), 201);
        } catch (InvalidCredentialsException $e) {
            return response()->json(['message' => $e->getMessage()], 401);
        }
    }

    /**
     * @OA\Post(
     *     path="/auth/logout",
     *     tags={"Auth"},
     *     summary="Close session",
     *     security={{"bearerAuth": {}}},
     *     @OA\Response(response=200, description="OK", @OA\JsonContent(ref="#/components/schemas/MessageResponse")),
     *     @OA\Response(response=401, description="Unauthorized", @OA\JsonContent(ref="#/components/schemas/ErrorResponse")),
     *     @OA\Response(response=500, description="Server error", @OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function logout(): JsonResponse
    {
        try {
            $this->logout->execute();
            return response()->json(['message' => 'Successfully logged out'], 200);
        } catch (Exception $e) {
            return response()->json(['message' => 'Error logging out'], 500);
        }
    }

    /**
     * @OA\Post(
     *     path="/auth/refresh",
     *     tags={"Auth"},
     *     summary="Refresh token",
     *     security={{"bearerAuth": {}}},
     *     @OA\Response(response=200, description="OK", @OA\JsonContent(ref="#/components/schemas/AuthResponse")),
     *     @OA\Response(response=401, description="Unauthorized", @OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function refresh(): JsonResponse
    {
        $response = $this->refreshToken->execute();
        return response()->json($response->toArray(), 200);
    }

    /**
     * @OA\Get(
     *     path="/auth/me",
     *     tags={"Auth"},
     *     summary="Get authenticated user data",
     *     security={{"bearerAuth": {}}},
     *     @OA\Response(response=200, description="OK", @OA\JsonContent(ref="#/components/schemas/ProfileResponse")),
     *     @OA\Response(response=401, description="Unauthorized", @OA\JsonContent(ref="#/components/schemas/ErrorResponse"))
     * )
     */
    public function me(): JsonResponse
    {
        $userData = $this->me->execute();
        return response()->json($userData, 200);
    }
}